Cleaned states and actions
